refactor(live-window): move sky-gradient into layers/gradient with SkyLayer class
